Roll ondemand code into spotbid
